+++ cockatoo/__init__.py
@@ -2,7 +2,7 @@
 # :copyright: (c) 2013 Andrew E. Bruno
 # :license:   GPL v3, see LICENSE for more details.
 
-VERSION = (0, 6, 0)
+VERSION = (0, 6, 1)
 __version__ = ".".join(map(str, VERSION[:]))
 
-from cockatoo import screen,metric
+from cockatoo import screen,metric,xtuition